[clean] Created OtherUserMessageBuilder. This component display other users messages.
