+++ chat-frontend/src/pages/chat/component/activeChat.jsx
@@ -8,6 +8,7 @@ import { useActiveChatDataContex } from "../../../context/activeChatDataContext"
 import { useChatDataContex } from "../../../context/chatContext";
 import ActiveChatInput from "./activeChatComponent/activeChatInputField";
 import OwenMessageBuiler from "./activeChatComponent/ownMessageBuilder";
+import OtherUserMessageBuilder from "./activeChatComponent/otherUserMessageBuilder";
 
 const ActiveChat = () => {
 
@@ -16,11 +17,7 @@ const ActiveChat = () => {
     const {value, index} = useActiveChatDataContex();
     
 
-    const emojiClickHandler = (messageId) => {
-
-        useStateSetObject.setClickEmojiPicker(messageId);
-
-    }
+    
 
     const splitMessageHistory = async (currentIndex) => {
 
@@ -133,58 +130,7 @@ const ActiveChat = () => {
                                 if (message.UserId === useStateValueObject.profileData.id) {
                                     return <OwenMessageBuiler message={message} index={index}/>
                                 } else {
-                                    return <div className="chatMessageWrapper"
-                                        id={message.ChatId + "." + index}
-                                        onMouseEnter={() => { useStateSetObject.setOnFocusMessage(message.MessageId) }}
-                                        onMouseLeave={() => { useStateSetObject.setOnFocusMessage(null), useStateSetObject.setClickEmojiPicker(null) }}>
-                                        <div className="messageContentContainer">
-
-                                            <img className="chatuserIcon" src={User} alt="chat user Icon" />
-
-                                            <div className="messageContainer">
-                                                <div key={index} className="chatRecivedMessage">
-                                                    {message.Content}
-                                                    <div className="chatEmoji"
-                                                        style={{
-                                                            visibility: message.Emoji.length === 0 ? 'hidden' : 'visible',
-                                                        }}>
-                                                        {
-                                                            message.Emoji.map((value, index) => {
-                                                                if (index < 3) {
-                                                                    return <DisplayEmoji emojiValue={value}/>
-                                                                }
-                                                            })
-                                                        }
-                                                    </div>
-
-                                                </div>
-
-
-                                                <div className="addEmoji"
-                                                    style={{
-                                                        visibility: useStateValueObject.onFocusMessage === message.MessageId ? 'visible' : 'hidden',
-                                                    }}
-                                                    onClick={() => { emojiClickHandler(message.MessageId) }}>
-                                                    {
-                                                        useStateValueObject.clickEmojiPicker === message.MessageId ? (
-
-                                                            <Emoji  chatId={message.ChatId} messageId={message.MessageId}/>
-
-                                                        ) : (
-                                                            <></>
-                                                        )
-                                                    }
-
-                                                    <img className="addEmojiIcon" src={Happiness} alt="chat add emoji Icon" />
-
-                                                </div>
-
-                                            </div>
-
-
-                                        </div>
-
-                                    </div>
+                                    return <OtherUserMessageBuilder message={message} index={index} />
                                 } 
                             }
